refactor: 전시회 푸시 알림 요청 사용자 식별 방식 token에서 닉네임으로 변경
