+++ src/main/java/dandi/dandi/exhibition/ExhibitionPushNotificationRequest.java
@@ -1,0 +1,13 @@
+package dandi.dandi.exhibition;
+
+public class ExhibitionPushNotificationRequest {
+
+    private String nickname;
+
+    public ExhibitionPushNotificationRequest() {
+    }
+
+    public String getNickname() {
+        return nickname;
+    }
+}
